changing namespaces for spliting benchmarks up to organize better
