+++ src/ef-json-query-testing/Enums/DataTypes.cs
@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ef_json_query_testing
+﻿namespace ef_json_query_testing.Enums
 {
     public enum DataTypes
     {
