+++ src/ef-json-query-testing/Models/DynamicListItem.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace ef_json_query_testing
+namespace ef_json_query_testing.Models
 {
     public class DynamicListItem
     {
