+++ src/ef-json-query-testing/Models/DynamicListType.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ef_json_query_testing
+﻿namespace ef_json_query_testing.Models
 {
     public class DynamicListType
     {
