+++ src/ef-json-query-testing/Models/Media_Dynamic.cs
@@ -1,13 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Text.Json;
-using System.Threading.Tasks;
-
-namespace ef_json_query_testing
+﻿namespace ef_json_query_testing.Models
 {
     public class Media_Dynamic
     {
